fixed setting values in context for additionalvalues
